pyproject.toml
[tool.poetry]
name = "iis"
version = "0.1.0"
description = ""
authors = ["PolnerFilip <filip.polner@student.um.si>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.13"
pandas = "^2.2.1"
requests = "^2.31.0"
schedule = "^1.2.1"
jupyter = "^1.0.0"
colorlog = "^6.8.2"
tensorflow = "2.10.0"
tensorflow-io-gcs-filesystem = "0.27.0"
python-dotenv = "^1.0.1"
scikit-learn = "^1.4.1.post1"
ipykernel = "^6.29.3"
flask = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
